tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-auto-tracks-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masonry: masonry-auto-tracks with valid values</title>
<link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-3">
<meta name="assert" content="masonry-auto-tracks supports the full grammar '<track-size>+'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// <track-breadth>
// <track-breadth> = <length-percentage> | <flex> | min-content | max-content | auto
test_valid_value("masonry-auto-tracks", "1px");
test_valid_value("masonry-auto-tracks", "2em");
test_valid_value("masonry-auto-tracks", "calc(2em + 3ex)");
test_valid_value("masonry-auto-tracks", "4%");
test_valid_value("masonry-auto-tracks", "5fr");
test_valid_value("masonry-auto-tracks", "min-content");
test_valid_value("masonry-auto-tracks", "max-content");
test_valid_value("masonry-auto-tracks", "auto");
test_valid_value("masonry-auto-tracks", "auto /**/", "auto");

// minmax( <inflexible-breadth> , <track-breadth> )
// <inflexible-breadth> = <length-percentage> | min-content | max-content | auto
test_valid_value("masonry-auto-tracks", "minmax(1px, 5fr)");
test_valid_value("masonry-auto-tracks", "minmax(2em, min-content)");
test_valid_value("masonry-auto-tracks", "minmax(calc(2em + 3ex), max-content)");
test_valid_value("masonry-auto-tracks", "minmax(4%, auto)");
test_valid_value("masonry-auto-tracks", "minmax(5vmin, 1px)");
test_valid_value("masonry-auto-tracks", "minmax(min-content, 2em)");
test_valid_value("masonry-auto-tracks", "minmax(max-content, calc(2em + 3ex))");
test_valid_value("masonry-auto-tracks", "minmax(auto, 4%)");

// fit-content( <length-percentage> )
test_valid_value("masonry-auto-tracks", "fit-content(1px)");
test_valid_value("masonry-auto-tracks", "fit-content(2em)");
test_valid_value("masonry-auto-tracks", "fit-content(calc(2em + 3ex))");
test_valid_value("masonry-auto-tracks", "fit-content(4%)");

test_valid_value("masonry-auto-tracks", "0px");
test_valid_value("masonry-auto-tracks", "0%");
test_valid_value("masonry-auto-tracks", "0fr");
test_valid_value("masonry-auto-tracks", "minmax(auto, 0%)");
test_valid_value("masonry-auto-tracks", "fit-content(0px)");

// <track-size>+
test_valid_value("masonry-auto-tracks", "auto auto");
test_valid_value("masonry-auto-tracks", "auto 10px");
test_valid_value("masonry-auto-tracks", "1px 2px 3px 0px");
test_valid_value("masonry-auto-tracks", "fit-content(1px) minmax(2px, 3px) 4px");
</script>
</body>
</html>
